requirements.txt
ollama
mcp
python-dotenv
rich
prompt_toolkit
httpx
